level 11 - Done code refactoring according to SOLID principles
